Made the NodeComparator class. Finished the Astar, not really Astar, some searching algo. Need to change that name. Have not tested yet.

+++ Frame/Coordinate.java
@@ -98,7 +98,7 @@ public class Coordinate {
 		return "(" + x + ", " + y + ")";
 	}
 	
-	public Node spawnNode(int aPriority) {
+	public Node spawnNode(double aPriority) {
 		return new Node(aPriority, this);
 	}
 }
